progress on tableProposals
