+++ app/src/renderer/components/governance/TableProposals.vue
@@ -1,0 +1,170 @@
+<template lang="pug">
+div
+  tm-data-loading(v-if="loading")
+  tm-data-empty(v-else-if="proposals.length === 0")
+  data-empty-search(v-else-if="filteredProposals.length === 0")
+  table(v-else)
+    thead
+      panel-sort(:sort='sort', :properties="properties")
+    tbody
+      li-proposal(
+        v-for="(value, key) in filteredProposals"
+           :key="key"
+           :proposal="value")
+</template>
+
+<script>
+import { mapGetters } from "vuex"
+import { includes, orderBy } from "lodash"
+import LiProposal from "./LiProposal"
+import { TmDataEmpty, TmDataLoading } from "@tendermint/ui"
+import DataEmptySearch from "common/TmDataEmptySearch"
+import { ratToBigNumber } from "scripts/common"
+import ModalSearch from "common/TmModalSearch"
+import PanelSort from "staking/PanelSort"
+import VmToolBar from "common/VmToolBar"
+export default {
+  name: `table-proposals`,
+  props: [`proposals`, `loading`],
+  components: {
+    LiProposal,
+    TmDataEmpty,
+    DataEmptySearch,
+    TmDataLoading,
+    ModalSearch,
+    PanelSort,
+    VmToolBar
+  },
+  data: () => ({
+    query: ``,
+    sort: {
+      property: `proposal_id`,
+      order: `desc`
+    }
+  }),
+  computed: {
+    ...mapGetters([`filters`, `config`]),
+    // TODO delete once tally is changed from Rat --> Dec
+    parsedProposals() {
+      return !this.proposals || this.proposals.length === 0
+        ? []
+        : Object.values(this.proposals).map(p => {
+            p.tally_result.yes = Math.round(
+              ratToBigNumber(p.tally_result.yes).toNumber()
+            )
+            p.tally_result.no = Math.round(
+              ratToBigNumber(p.tally_result.no).toNumber()
+            )
+            p.tally_result.no_with_veto = Math.round(
+              ratToBigNumber(p.tally_result.no_with_veto).toNumber()
+            )
+            p.tally_result.abstain = Math.round(
+              ratToBigNumber(p.tally_result.abstain).toNumber()
+            )
+            return p
+          })
+    },
+    filteredProposals() {
+      if (this.proposals && this.filters) {
+        let query = this.filters.proposals.search.query
+        let proposals = orderBy(
+          this.parsedProposals,
+          [this.sort.property],
+          [this.sort.order]
+        )
+        if (this.filters.proposals.search.visible) {
+          return proposals.filter(p =>
+            includes(p.title.toLowerCase(), query.toLowerCase())
+          )
+        } else {
+          return proposals
+        }
+      } else {
+        return []
+      }
+    },
+    properties() {
+      return [
+        {
+          title: `Proposal ID`,
+          value: `proposal_id`,
+          tooltip: `The ID of the proposal`,
+          class: `proposal_id`
+        },
+        {
+          title: `Yes`,
+          value: `tally_result.yes`,
+          tooltip: `Yes votes`,
+          class: `yes_votes`
+        },
+        {
+          title: `No`,
+          value: `tally_result.no`,
+          tooltip: `No votes`,
+          class: `no_votes`
+        },
+        {
+          title: `No with Veto`,
+          value: `tally_result.no_with_veto`,
+          tooltip: `No with veto votes`,
+          class: `no_veto_votes`
+        },
+        {
+          title: `Abstain`,
+          value: `tally_result.abstain`,
+          tooltip: `Abstain votes`,
+          class: `abstain_votes`
+        }
+      ]
+    }
+  }
+}
+</script>
+<style lang="stylus">
+@require '~variables'
+
+table
+  border-spacing 0 0.25rem
+  margin 0 0 0 2rem
+  min-width
+  padding 0
+  table-layout auto
+  counter-reset rowNumber + 1
+
+table tr
+  counter-increment rowNumber
+
+table tr td:first-child::before
+  content counter(rowNumber)
+  position absolute
+  font-size sm
+  width 2rem
+  text-align right
+  color var(--dim)
+  left -3rem
+
+table th
+  min-width 130px
+  width 100%
+  padding 0.5rem
+
+table td
+  min-width 130px
+  width 100%
+  padding 0 0.5rem
+  position relative
+
+  a
+    display inline-block
+
+table tr td:nth-child(3):after
+  display block
+  position absolute
+  content ''
+  height 2rem
+  width 2px
+  top 1.5rem
+  right 2rem
+  background var(--bc-dim)
+
+</style>
