+++ app/src/renderer/components/staking/TabValidators.vue
@@ -7,6 +7,7 @@ import TableValidators from "staking/TableValidators"
 import { mapGetters } from "vuex"
 
 export default {
+  name: `tab-validators`,
   components: {
     TableValidators
   },
